knx: widen the MMIO/BAR ABI to 64-bit — unblocks real-HW GPUs (Dell Comet Lake) with BARs above 4 GiB
The Phase-B blocker flagged in kpi_iomap.c: knx_map_mmio / knx_pci_bar were 32-bit, so a device
whose firmware places a BAR above 4 GiB — e.g. the Dell's Comet Lake GTTMMADR/GTT — truncated to
the low 32 bits and would fault on the first register access. The kernel MMU already identity-maps
any physical address (mmuMapKernelMmio takes uint64_t phys and builds page tables for it), so this
is purely an ABI/decode widening; nothing below the shim needed new capability.

Kernel:
  - PciBar.addr/size uint32_t -> uint64_t; Pci::readBars now combines the high dword of a 64-bit
    BAR into both the base and the size mask (32-bit BARs treat the absent high bits as all-ones so
    ~mask+1 stays correct; a both-dwords-zero read-back is the unimplemented/size-0 sentinel).
  - knx_pci_bar / knx_pci_bar_size return uint64_t; knx_map_mmio(uint64_t phys, uint64_t len);
    mmuMapKernelMmio(uint64_t phys, uint64_t bytes) (both arches).
  - MsiRouter: the MSI-X table BAR base is read as 64-bit (folds in the high dword when the BAR is
    64-bit — e.g. i915's MSI-X lives in BAR0), and MsiEnv.mapMmio widened to match.

Shim: dropped the truncating (unsigned) casts in io.h (ioremap*/devm_ioremap*/memremap),
device.h (devm_request_mem_region), kpi_pci.c (pci_iomap_range), kpi_iomap.c (io_mapping); the
64-bit phys_addr_t / resource base now passes through intact. lkpi_knx.h + e1000 extern updated.

Verified: test64 863/863 (test_pci exercises both a 32-bit and a 64-bit BAR decode; test_msi_router
covers the MSI-X BAR path), and the widened path is unregressed at runtime — virtio_gpu probes and
bridges /dev/fb0, e1000e brings eth0 up, and smoke-kpi-irq delivers a real MSI-X interrupt, all on
their existing sub-4 GiB BARs.

+++ arch/include/arch/mmu.h
@@ -52,7 +52,7 @@ uint32_t mmuKernelDirPhys();
 // routinely programmed above 4 GiB. Truncating to 32 bits maps the wrong page and the
 // first access triple-faults on the metal (QEMU/OVMF places them low, hiding it). The
 // i686 impl narrows internally (its physical space is 32-bit anyway).
-void mmuMapKernelMmio(uint64_t phys, uint32_t bytes);
+void mmuMapKernelMmio(uint64_t phys, uint64_t bytes);
 
 // Read/load the active page-directory physical address (CR3 on x86). MI code uses
 // these to stage a spawned child under the kernel identity map, then restore the
